Read a Socrata Dataset into R

Downloads and parses a dataset from a Socrata open data portal URL, returning it as a tibble or sf object. Metadata is also returned as attributes on the returned object.

Usage

soc_read(url, query = soc_query(), alias = "label", page_size = 10000)

Arguments

url

string; URL of the Socrata dataset.

query

soc_query(); Query parameters specification

alias

string; Use of field alias values. There are three options:

"label": field alias values are assigned as a label attribute for each field.

"replace": field alias values replace existing column names.

"drop": field alias values replace existing column names.

page_size

whole number; Maximum number of rows returned per request.

Value

A tibble with additional attributes containing dataset metadata. If the dataset contains a single non-nested geospatial field, it will be returned as an sf object.

The returned object has the following attributes:

id

Asset identifier (four-by-four ID).

name

Asset name.

attribution

Attribution or publisher of the asset.

owner_name

Display name of the asset owner.

provenance

Provenance of asset (official or community).

description

Textual description of the asset.

created

Date asset was created.

data_last_updated

Date asset data was last updated

metadata_last_updated

Date asset metadata was last updated

domain_category

Category label assigned by the domain.

domain_tags

Tags applied by the domain.

domain_metadata

Metadata associated with the asset assigned by the domain.

columns

A dataframe with the following columns:

column_name

Names of asset columns.

column_label

Labels of asset columns.

column_datatype

Datatypes of asset columns.

column_description

Description of asset columns.

permalink

Permanent URL where the asset can be accessed.

link

Direct asset link.

license

License associated with the asset.

Examples

# \donttest{
cta_ridership <- soc_read(
  "https://data.cityofchicago.org/Transportation/Speed-Camera-Violations/hhkd-xvj4/about_data"
)
#> iterating ■■                                 2% | ETA: 44s
#> iterating ■■                                 5% | ETA:  1m
#> iterating ■■■■                               9% | ETA:  1m
#> iterating ■■■■■■                            16% | ETA: 49s
#> iterating ■■■■■■                            18% | ETA:  1m
#> iterating ■■■■■■■■                          23% | ETA:  1m
#> iterating ■■■■■■■■■                         27% | ETA: 46s
#> iterating ■■■■■■■■■■■                       32% | ETA: 43s
#> iterating ■■■■■■■■■■■■                      36% | ETA: 42s
#> iterating ■■■■■■■■■■■■■                     41% | ETA: 41s
#> iterating ■■■■■■■■■■■■■■                    43% | ETA: 39s
#> iterating ■■■■■■■■■■■■■■■                   45% | ETA: 39s
#> iterating ■■■■■■■■■■■■■■■■                  50% | ETA: 36s
#> iterating ■■■■■■■■■■■■■■■■■                 55% | ETA: 33s
#> iterating ■■■■■■■■■■■■■■■■■■                57% | ETA: 32s
#> iterating ■■■■■■■■■■■■■■■■■■■               59% | ETA: 31s
#> iterating ■■■■■■■■■■■■■■■■■■■■              64% | ETA: 27s
#> iterating ■■■■■■■■■■■■■■■■■■■■■             68% | ETA: 24s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■            70% | ETA: 22s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■           73% | ETA: 21s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■          77% | ETA: 18s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■        82% | ETA: 14s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■        84% | ETA: 13s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86% | ETA: 11s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      91% | ETA:  7s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     93% | ETA:  5s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    98% | ETA:  2s
#> iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
print(cta_ridership)
#> # A tibble: 434,000 × 14
#>    address    camera_id violation_date      violations x_coordinate y_coordinate
#>    <chr>      <chr>     <dttm>                   <dbl>        <dbl>        <dbl>
#>  1 10318 S I… CHI120    2025-06-15 00:00:00        126     1203645.     1837056.
#>  2 10540 S W… CHI205    2025-06-15 00:00:00        525     1162222.     1834494.
#>  3 1111 N HU… CHI010    2025-06-15 00:00:00        116          NA           NA 
#>  4 1142 W IR… CHI095    2025-06-15 00:00:00         91     1167790.     1926747.
#>  5 115 N OGD… CHI077    2025-06-15 00:00:00         36     1166485.     1900735.
#>  6 1274 E 83… CHI197    2025-06-15 00:00:00        120     1186302.     1850256.
#>  7 1306 W 76… CHI185    2025-06-15 00:00:00         46     1168833.     1854443.
#>  8 1315 W GA… CHI121    2025-06-15 00:00:00         83     1168445.     1868118.
#>  9 1334 W GA… CHI119    2025-06-15 00:00:00         95     1168211.     1868328.
#> 10 1461 W GR… CHI208    2025-06-15 00:00:00        150     1166314.     1903542.
#> # ℹ 433,990 more rows
#> # ℹ 8 more variables: latitude <dbl>, longitude <dbl>, location <tibble[,5]>,
#> #   `:@computed_region_awaf_s7ux` <dbl>, `:@computed_region_6mkv_f3dw` <dbl>,
#> #   `:@computed_region_vrxf_vc4k` <dbl>, `:@computed_region_bdys_3d7i` <dbl>,
#> #   `:@computed_region_43wa_7qmu` <dbl>
attr(cta_ridership, "description")
#> [1] "This dataset reflects the daily volume of violations that have occurred in Children's Safety Zones for each camera. The data reflects violations that occurred from July 1, 2014 until present, minus the most recent 14 days. This data may change due to occasional time lags between the capturing of a potential violation and the processing and determination of a violation. The most recent 14 days are not shown due to revised data being submitted to the City of Chicago. The reported violations are those that have been collected by the camera and radar system and reviewed by two separate City contractors. In some instances, due to the inability the registered owner of the offending vehicle, the violation may not be issued as a citation. However, this dataset contains all violations regardless of whether a citation was issued, which provides an accurate view into the Automated Speed Enforcement Program violations taking place in Children's Safety Zones. More information on the Safety Zone Program can be found here: http://www.cityofchicago.org/city/en/depts/cdot/supp_info/children_s_safetyzoneporgramautomaticspeedenforcement.html. The corresponding dataset for red light camera violations is https://data.cityofchicago.org/id/spqx-js37."

trips_to_lws_by_ca <- soc_read(
  "https://data.cityofchicago.org/transportation/taxi-trips-2013-2023-/wrvz-psew/about_data",
  query = soc_query(
    select = "violation_date, count(*) as n",
    where = "dropoff_community_area = 31",
    group_by = "pickup_community_area",
    order_by = "n DESC"
  ),
  alias = "replace"
)
#> Error in req_perform(req, path = get_path(i), mock = mock): HTTP 400 Bad Request.
# }