Read a Socrata Dataset into R

Downloads and parses a dataset from a Socrata open data portal URL, returning it as a tibble or sf object. Metadata is also returned as attributes on the returned object.

Usage

soc_read(url, query = soc_query(), alias = "label")

Arguments

url

string; the URL of the Socrata dataset (e.g., from https://data.cityofchicago.org).

query

soc_query(); query parameters specification

alias

string; use of field alias values. There are two options:

"label": field alias values are assigned as a label attribute for each field.

"replace": field alias values replace existing column names.

Value

A socrata_tbl, which is a tibble with additional class and attributes containing dataset metadata. If the dataset contains a single non-nested geospatial field, it will be returned as an sf object.

The returned object has the following attributes:

id

Dataset identifier (four-by-four ID).

name

Dataset name.

attribution

Attribution or publisher of the dataset.

category

Category label assigned on the portal.

created

POSIXct timestamp when the dataset was created.

data_last_updated

POSIXct timestamp of the last data update.

metadata_last_updated

POSIXct timestamp of the last metadata update.

description

Textual description of the dataset.

Examples

if (FALSE) { # \dontrun{
cta_ridership <- soc_read(
  "https://data.cityofchicago.org/Transportation/CTA-Ridership-Daily-Boarding-Totals/6iiy-9s97/about_data"
)
print(cta_ridership)
attr(cta_ridership, "description")

trips_to_lws_by_ca <- soc_read(
  "https://data.cityofchicago.org/Transportation/Taxi-Trips-2013-2023-/wrvz-psew/about_data",
  query = soc_query(
    select = "pickup_community_area, count(*) as n",
    where = "dropoff_community_area = 31",
    group_by = "pickup_community_area",
    order_by = "n DESC"
  ),
  alias = "replace"
)
} # }